pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "plasflow-v2"
version = "2.0.0a0"
description = "PlasFlow v2: modernized plasmid/chromosome classifier with CLI, API, and web UI"
readme = "README_v2.md"
requires-python = ">=3.11"
license = {text = "GPL-3.0-or-later"}
authors = [
  {name = "PlasFlow Contributors"}
]
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-multipart>=0.0.9",
  "joblib>=1.4"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "httpx>=0.27"
]
train = [
  "scikit-learn>=1.5",
  "lightgbm>=4.5",
  "pyyaml>=6.0.2"
]

[project.scripts]
plasflow = "plasflow_v2.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/plasflow_v2"]
